chore(src): rename `components` dir
`conponents`->`components`
