+++ src/components/BrowserLocationQuery/index.ts
